refactor: standarize the use of usize for unsigned integers
usize is used in a lot of places in the code by cast os as u32. To
standarize the usage of unsigned integers, usize is going to be used
in detriment of u32.

Closes #35

+++ src/app/config.rs
@@ -1,7 +1,7 @@
 use std::env;
 
 pub struct Config {
-    pub page_size: u32,
+    pub page_size: usize,
     pub patchsets_cache_dir: String,
     pub bookmarked_patchsets_path: String,
     pub mailing_lists_path: String,
@@ -10,7 +10,7 @@ pub struct Config {
 
 impl Config {
     pub fn build() -> Self {
-        let page_size: u32 = match env::var("PATCH_HUB_PAGE_SIZE") {
+        let page_size: usize = match env::var("PATCH_HUB_PAGE_SIZE") {
             Ok(value) => value.parse().unwrap(),
             Err(_) => 30,
         };
